refactor: move CoreObjectSerializer to Serializable
