+++ src/Server/Game/Services/AttachService.cs
@@ -1,4 +1,5 @@
 ﻿using Akka.Actor;
+using Imlight.Common.Serializable;
 using WizUnraveler;
 using WizUnraveler.Cache;
 using static WizUnraveler.Cache.TypeCache;
