+++ src/Server/Game/Zone.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Akka.Actor;
+using Imlight.Common.Serializable;
 using WizUnraveler;
 using WizUnraveler.Cache;
 using WizUnraveler.DML;
